+++ src/main/java/org/example/model/Product.java
@@ -2,6 +2,7 @@ package org.example.model;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 
 
 @Entity
@@ -18,12 +19,12 @@ public class Product  {
     @Column(name = "description", nullable = false)
     private String description;
     @Column(name = "price", nullable = false)
-    private double price;
+    private BigDecimal price;
 
     public Product() {
     }
 
-    public Product(Long id, String name, String description, double price) {
+    public Product(Long id, String name, String description, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -56,11 +57,11 @@ public class Product  {
         this.description = description;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
